demos/agent-systems-playground/requirements.txt
streamlit>=1.32.0
openai>=1.12.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
